pyproject.toml
[project]
name = "VT1"
version = "0.1.0"
description = "Visualization of Landscape Changes using Synthesized Aerial Imagery"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.3.0",
    "datasets>=3.2.0",
    "diffusers>=0.32.2",
    "imageio>=2.37.0",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.0",
    "numpy>=2.2.2",
    "omegaconf>=2.3.0",
    "openpyxl>=3.1.5",
    "pillow>=11.1.0",
    "rasterio>=1.4.3",
    "tensorboard>=2.18.0",
    "torch-tb-profiler>=0.4.3",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "tqdm>=4.67.1",
    "transformers>=4.48.2",
]
